django/requirements.txt
django
djangorestframework
whitenoise
django-oauth-toolkit
gunicorn
